+++ ICD.Connect.Audio.Biamp/Controls/Dialing/Telephone/TiConferenceDeviceControl.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using ICD.Common.Properties;
 using ICD.Common.Utils.EventArguments;
 using ICD.Common.Utils;
 using ICD.Common.Utils.Extensions;
@@ -26,6 +27,8 @@ namespace ICD.Connect.Audio.Biamp.Controls.Dialing.Telephone
 		public override event EventHandler<GenericEventArgs<IIncomingCall>> OnIncomingCallRemoved;
 
 		private readonly TiControlStatusBlock m_TiControl;
+
+		[CanBeNull]
 		private readonly IBiampTesiraStateDeviceControl m_HoldControl;
 
 		private bool m_Hold;
@@ -580,7 +583,7 @@ namespace ICD.Connect.Audio.Biamp.Controls.Dialing.Telephone
 			if (holdControl == null)
 				return;
 
-			m_HoldControl.OnStateChanged += HoldControlOnStateChanged;
+			holdControl.OnStateChanged += HoldControlOnStateChanged;
 		}
 
 		/// <summary>
@@ -592,7 +595,7 @@ namespace ICD.Connect.Audio.Biamp.Controls.Dialing.Telephone
 			if (holdControl == null)
 				return;
 
-			m_HoldControl.OnStateChanged -= HoldControlOnStateChanged;
+			holdControl.OnStateChanged -= HoldControlOnStateChanged;
 		}
 
 		/// <summary>