+++ CHANGELOG.md
@@ -40,10 +40,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - Added port configuration features to audio devices
 
+## [9.3.0] - 2019-07-26
+### Added
+ - Added Shure MXWAPT4 support
+
 ## [9.2.0] - 2019-07-16
 ### Changed
  - Better handling of persistent connector info for telemetry
  - PowerOn/PowerOff methods changed to support new pre-on/off callbacks
+
+## [9.1.1] - 2019-07-26
+### Changed
+ - Fixed null reference issues when a Biamp TI control is instantiated without a hold control
 
 ## [9.1.0] - 2019-05-15
 ### Added